move `shared/commonProtocol` to `libs/evm-protocols`
