+++ libs/evm-protocols/src/index.ts
@@ -8,5 +8,5 @@ export * from './abis/namespaceAbi';
 export * from './abis/namespaceFactoryAbi';
 export * from './abis/reservationHookAbi';
 export * from './abis/tokenCommunityManagerAbi';
-export * as commonProtocol from './commonProtocol';
+export * as commonProtocol from './common-protocol';
 export * from './event-registry/eventSignatures';
