+++ libs/evm-protocols/test/commonProtocol.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { calculateVoteWeight } from '../src/commonProtocol';
+import { calculateVoteWeight } from '../src/common-protocol';
 
 describe('commonProtocol', () => {
   describe('calculateVoteWeight', () => {
